@@ -16,10 +16,10 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
 # In Windows, build-name is used as the major, minor, and patch parts
 # of the product and file versions while build-number is used as the build suffix.
-version: 1.0.4+9
+version: 1.0.5+10
 
 environment:
-  sdk: '>=3.3.3 <4.0.0'
+  sdk: '>=3.5.0 <4.0.0'
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -34,7 +34,7 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.8
-  intl: ^0.18.1
+  intl: ^0.19.0
   collection: ^1.17.0
   get: ^4.6.6
   dio: ^5.6.0
@@ -91,14 +91,22 @@ dependencies:
   html: ^0.15.4
   tutorial_coach_mark: ^1.2.11
   pal_widgets: ^0.3.0
-  flutter_inappwebview: ^6.0.0
+  flutter_inappwebview: ^6.1.1
   app_links: ^6.2.0
   dartz: ^0.10.1
   in_app_review: ^2.0.9
   easy_debounce: ^2.0.3
-  
+  firebase_messaging: ^15.1.0
+  basic_utils: ^5.7.0
+  dart_jsonwebtoken: ^2.14.0
+  pointycastle: ^3.9.1
+  flutter_cache_manager: ^3.4.1
+
 dependency_overrides:
   webview_flutter_android: 3.16.1
+  # fix build xcode 16 -> flutter_inappwebview
+  # url_launcher_ios: 6.3.1
+  
 
 dev_dependencies:
   flutter_launcher_icons: ^0.13.1
@@ -129,8 +137,8 @@ flutter:
     - assets/images/
     - assets/backgrounds/
     - assets/rive/
-    - assets/rive/plants/
-    - assets/rive/pots/
+    # - assets/rive/plants/
+    # - assets/rive/pots/
     - assets/rive/overlay/
     - assets/audios/
 
