@@ -188,5 +188,20 @@ Map<String, String> ru = {
   "(Bật)": "(На)",
   "Thiết lập": "Учреждать",
   "Tăng 1 cấp": "Повышение на 1 уровень",
-  "Tăng 1 cấp độ người dùng": "Повысить уровень пользователя на 1"
+  "Tăng 1 cấp độ người dùng": "Повысить уровень пользователя на 1",
+  "Kim tiền thân vàng": "Металл – предшественник золота",
+  "Loài cây có những nhánh lá xếp với nhau, cây có sức sống tốt và là loài có độc":
+      "У дерева ветви и листья расположены вместе, дерево обладает хорошей жизнеспособностью и ядовито.",
+  "Khoai nước thân vàng": "Водяной картофель с желтым телом",
+  "Loài khoai không có củ, lá to, cuống vàng rất đẹp":
+      "У этого вида картофеля нет клубней, крупные листья и красивые желтые стебли.",
+  "Cây nắp ấm đỏ": "Красный кувшин",
+  "Loài cây leo, phần đầu cuốn nổi bật với tạo hình giống như cái bình nước hình trụ":
+      "Вьющееся растение со свернутой верхушкой, по форме напоминающее цилиндрическую бутылку с водой.",
+  "Cây bẫy kẹp": "Зажимная ловушка",
+  "Cây bẫy kẹp là một loài thực vật ăn thịt, sống bản địa tại vùng đất ngập nước cận nhiệt đới":
+      "Растение-ловушка — хищное растение, произрастающее в субтропических водно-болотных угодьях.",
+  "Chậu đất treo": "Подвесные глиняные горшки",
+  "Chậu chậu đất thoát nước tốt, nhưng hơi nặng khi treo":
+      "Почвенные горшки хорошо дренируют воду, но их тяжело подвешивать."
 };
