@@ -1,6 +1,8 @@
+
 import 'dart:io';
 import 'package:firebase_analytics/firebase_analytics.dart';
 import 'package:firebase_core/firebase_core.dart';
+import 'package:firebase_messaging/firebase_messaging.dart';
 import 'package:flutter_phoenix/flutter_phoenix.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
@@ -12,10 +14,11 @@ import 'package:terrarium_idle/c_theme/c_theme.dart';
 import 'package:terrarium_idle/config/config.dart';
 import 'package:terrarium_idle/config/get_config.dart';
 import 'package:terrarium_idle/firebase_options.dart';
-import 'package:terrarium_idle/function/app_links_service.dart';
+import 'package:terrarium_idle/service/app_links_service.dart';
 import 'package:terrarium_idle/modules/init.dart';
 import 'package:terrarium_idle/modules/routers.dart';
 import 'package:terrarium_idle/modules/splash/splash_screen.dart';
+import 'package:terrarium_idle/service/firebase_push.dart';
 import 'package:terrarium_idle/service/local_notification.dart';
 
 Future<void> main() async {
@@ -27,14 +30,16 @@ Future<void> main() async {
   await Firebase.initializeApp(
     options: DefaultFirebaseOptions.currentPlatform,
   );
+  FirebaseMessaging.onBackgroundMessage(firebaseMessagingBackgroundHandler);
   // await JustAudioBackground.init(
   //   androidNotificationChannelId: 'com.ryanheise.bg_demo.channel.audio',
   //   androidNotificationChannelName: 'Audio playback',
   //   androidNotificationOngoing: true,
   // );
   await initialize();
-  initLocalNotification();
-  initDeepLinks();
+  LocalNotification().initLocalNotification();
+  AppLinksService().initDeepLinks();
+  FirebaseCouldMessage.init();
   runApp(
     Phoenix(
       child: const MyApp(),
